+++ packages/veui/src/components/Input.vue
@@ -97,7 +97,7 @@ export default {
   computed: {
     attrs () {
       return {
-        ...omit(this.$props, 'selectOnFocus', 'composition', 'value', 'clearable'),
+        ...omit(this.$props, 'placeholder', 'selectOnFocus', 'composition', 'value', 'clearable'),
         name: this.realName,
         disabled: this.realDisabled,
         readonly: this.realReadonly,